# Backend/docs/swagger.yaml
definitions:
...
  controllers.AdminStatsResponse:
    properties:
      akun_aktif:
        example: 142
        type: integer
      akun_inaktif:
        example: 8
        type: integer
      pending_device:
        example: 3
        type: integer
      total_akun:
        example: 150
        type: integer
    type: object
  controllers.CreateKelasReq:
    properties:
      nama_kelas:
        example: XII RPL 1
        type: string
    type: object
  controllers.CreateMapelReq:
    properties:
      nama_mapel:
        example: Pemrograman Web & Perangkat Bergerak
        type: string
    type: object
  controllers.DeviceActionReq:
    properties:
      device_id:
        example: 1
        type: integer
    type: object
...
  controllers.ResetPassSiswaReq:
    properties:
      kata_kunci:
        example: KucingKu
        type: string
      new_password:
        example: SandiBaru123
        type: string
      nis:
        example: "22010101"
        type: string
    type: object
...
  controllers.SuccessMessage:
    properties:
      message:
        example: Aksi berhasil dilakukan!
        type: string
    type: object
info:
  contact:
    email: hafizh@example.com
    name: Hafizh
  description: Ini adalah dokumentasi REST API untuk aplikasi absensi dengan fitur
    QR & Geofencing.
  termsOfService: http://swagger.io/terms/
  title: API Sistem Absensi Tempat Les
  version: "1.0"
paths:
  /admin/dashboard/stats:
    get:
      description: Menampilkan jumlah total akun, akun aktif (90 hari), inaktif, dan
        antrean device
      produces:
      - application/json
      responses:
        "200":
          description: Data Statistik
          schema:
            $ref: '#/definitions/controllers.AdminStatsResponse'
      security:
      - BearerAuth: []
      summary: Statistik Dashboard Admin
      tags:
      - 4. Admin - Dashboard & Device
  /admin/device/reject:
    post:
      consumes:
      - application/json
      description: Menolak akses login dari HP yang tidak dikenal
      parameters:
      - description: ID Perangkat
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/controllers.DeviceActionReq'
      produces:
      - application/json
      responses:
        "200":
          description: Perangkat Ditolak
          schema:
            $ref: '#/definitions/controllers.SuccessMessage'
      security:
      - BearerAuth: []
      summary: Tolak Perangkat Siswa
      tags:
      - 4. Admin - Dashboard & Device
  /admin/kelas/create:
    post:
      consumes:
      - application/json
      parameters:
      - description: Nama Kelas
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/controllers.CreateKelasReq'
      produces:
      - application/json
      responses:
        "200":
          description: Kelas berhasil ditambahkan
          schema:
            $ref: '#/definitions/controllers.SuccessMessage'
      security:
      - BearerAuth: []
      summary: Tambah Kelas Baru
      tags:
      - 5. Admin - Master Data
  /admin/mapel/create:
    post:
      consumes:
      - application/json
      parameters:
      - description: Nama Mapel
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/controllers.CreateMapelReq'
      produces:
      - application/json
      responses:
        "200":
          description: Mapel berhasil ditambahkan
          schema:
            $ref: '#/definitions/controllers.SuccessMessage'
      security:
      - BearerAuth: []
      summary: Tambah Mata Pelajaran Baru
      tags:
      - 5. Admin - Master Data
...
  /admin/siswa/reset-password:
    post:
      consumes:
      - application/json
      description: Admin mereset password siswa dengan validasi Kata Kunci Rahasia
      parameters:
      - description: Data Reset
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/controllers.ResetPassSiswaReq'
      produces:
      - application/json
      responses:
        "200":
          description: Password berhasil direset
          schema:
            $ref: '#/definitions/controllers.SuccessMessage'
      security:
      - BearerAuth: []
      summary: Reset Password Siswa (Lupa Password)
      tags:
      - 4. Admin - Device & Keamanan
...
securityDefinitions:
  BearerAuth:
    description: 'Masukkan token JWT dengan format: Bearer {token}'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"